chore(cli) Rename module
